pagination, delete check

+++ src/controllers/userPost.controller.ts
@@ -246,6 +246,7 @@ class _UserPostController {
       if (!description || !authorId || !userPostId)
         return res.status(errorCode.GENERIC).send({ message: errorMessage.MISSING_PARAMS });
       const created = await UserPostService.createOneComment({ description, authorId, userPostId });
+
       res.status(201).send({ message: successMessages.SUCCESS, data: created });
     } catch (error) {
       console.log("Error: ", error);
@@ -258,11 +259,15 @@ class _UserPostController {
   async delete(req: any, res: Response) {
     try {
       const { postId } = req.body;
+      const { id } = res.locals.user;
       if (!postId)
         return res.status(errorCode.GENERIC).send({ message: errorMessage.MISSING_PARAMS });
       const foundPost = await UserPostService.getOneUserPost({ id: postId });
       if (!foundPost)
         return res.status(errorCode.GENERIC).send({ message: errorMessage.NOT_FOUND });
+      if(    id !== foundPost.authorId){
+        return res.status(errorCode.GENERIC).send({ message: errorMessage.NOT_ALLOWED });
+      }
       await UserPostService.delete(postId);
       res.status(201).send({ message: successMessages.SUCCESS });
     } catch (error) {
